+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Adding HRID to the Inventory results list & show columns options. Refs UIIN-1262.
 * Display more detailed error message when updating ownership for holdings fails. Refs UIIN-3339.
 * ECS | Restore Held by Facet in Call number Browse. Refs UIIN-3414.
+* Use current tenant for MARC Holdings requests in View Source. Refs UIIN-3423.
 
 ## [13.0.4](https://github.com/folio-org/ui-inventory/tree/v13.0.4) (2025-04-18)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.3...v13.0.4)